requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.3
cssselect==0.9.1
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-mathfilters==0.4.0
django-session-timeout==0.0.3
djangorestframework==3.11.0
flake8==2.4.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
pep8==1.7.1
pyflakes==0.8.1
pylint==2.4.4
pytz==2018.9
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.7
wrapt==1.11.2
XlsxWriter==1.2.2
